optimize and organize code
