+++ src/app/settings/page.tsx
@@ -3,6 +3,7 @@
 import { useState } from "react";
 import { useSession, signOut } from "next-auth/react";
 import PageHeader from "@/components/PageHeader";
+import CreateUserModal from "@/components/CreateUserModal";
 import {
   User as UserIcon,
   Bell,
@@ -19,16 +20,6 @@ import {
   Users,
   Plus,
 } from "lucide-react";
-import { useCreateUser } from "@/lib/data-hooks/useCreateUser";
-import { type CreateUserRequest } from "@/lib/data-hooks/services/auth";
-import type { Role, User as PrismaUser } from "@prisma/client";
-
-interface CreatedUser {
-  user: PrismaUser & {
-    name: string;
-  };
-  defaultPassword: string;
-}
 
 const SettingsPage = () => {
   const { data: session } = useSession();
@@ -37,14 +28,6 @@ const SettingsPage = () => {
   const [showPassword, setShowPassword] = useState(false);
   const [isDeletingAccount, setIsDeletingAccount] = useState(false);
   const [showCreateUserModal, setShowCreateUserModal] = useState(false);
-  const [createdUser, setCreatedUser] = useState<CreatedUser | null>(null);
-
-  const {
-    createUser,
-    isLoading: isCreatingUser,
-    error: createUserError,
-    resetError,
-  } = useCreateUser();
 
   // Form states
   const [profileForm, setProfileForm] = useState({
@@ -57,13 +40,6 @@ const SettingsPage = () => {
     currentPassword: "",
     newPassword: "",
     confirmPassword: "",
-  });
-
-  const [createUserForm, setCreateUserForm] = useState<CreateUserRequest>({
-    firstName: "",
-    lastName: "",
-    email: "",
-    role: "MEMBER",
   });
 
   const [preferences, setPreferences] = useState({
@@ -111,31 +87,6 @@ const SettingsPage = () => {
 
   const handleSignOut = async () => {
     await signOut({ callbackUrl: "/sign-in" });
-  };
-
-  const handleCreateUser = async (e: React.FormEvent) => {
-    e.preventDefault();
-    resetError();
-
-    try {
-      const result = await createUser(createUserForm);
-      setCreatedUser(result);
-      setCreateUserForm({
-        firstName: "",
-        lastName: "",
-        email: "",
-        role: "MEMBER",
-      });
-      setShowCreateUserModal(false);
-    } catch (error) {
-      // Error is handled by the hook
-    }
-  };
-
-  const handleCloseCreateUserModal = () => {
-    setShowCreateUserModal(false);
-    setCreatedUser(null);
-    resetError();
   };
 
   return (
@@ -614,177 +565,10 @@ const SettingsPage = () => {
 
       {/* Create User Modal */}
       {showCreateUserModal && (
-        <div className="fixed inset-0 z-50 flex items-center justify-center bg-black bg-opacity-50">
-          <div className="mx-4 max-w-md rounded-lg bg-white p-6 shadow-xl">
-            <div className="mb-4 flex items-center justify-between">
-              <h3 className="text-lg font-medium text-gray-900">
-                Create New User
-              </h3>
-              <button
-                onClick={handleCloseCreateUserModal}
-                className="text-gray-400 hover:text-gray-600"
-              >
-                <X className="h-5 w-5" />
-              </button>
-            </div>
-
-            <form onSubmit={handleCreateUser} className="space-y-4">
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  First Name
-                </label>
-                <input
-                  type="text"
-                  value={createUserForm.firstName}
-                  onChange={(e) =>
-                    setCreateUserForm({
-                      ...createUserForm,
-                      firstName: e.target.value,
-                    })
-                  }
-                  required
-                  className="mt-1 block w-full rounded-lg border border-gray-300 px-3 py-2 shadow-sm focus:border-blue-500 focus:outline-none focus:ring-1 focus:ring-blue-500"
-                />
-              </div>
-
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Last Name
-                </label>
-                <input
-                  type="text"
-                  value={createUserForm.lastName}
-                  onChange={(e) =>
-                    setCreateUserForm({
-                      ...createUserForm,
-                      lastName: e.target.value,
-                    })
-                  }
-                  required
-                  className="mt-1 block w-full rounded-lg border border-gray-300 px-3 py-2 shadow-sm focus:border-blue-500 focus:outline-none focus:ring-1 focus:ring-blue-500"
-                />
-              </div>
-
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Email
-                </label>
-                <input
-                  type="email"
-                  value={createUserForm.email}
-                  onChange={(e) =>
-                    setCreateUserForm({
-                      ...createUserForm,
-                      email: e.target.value,
-                    })
-                  }
-                  required
-                  className="mt-1 block w-full rounded-lg border border-gray-300 px-3 py-2 shadow-sm focus:border-blue-500 focus:outline-none focus:ring-1 focus:ring-blue-500"
-                />
-              </div>
-
-              <div>
-                <label className="block text-sm font-medium text-gray-700">
-                  Role
-                </label>
-                <select
-                  value={createUserForm.role}
-                  onChange={(e) =>
-                    setCreateUserForm({
-                      ...createUserForm,
-                      role: e.target.value as "ADMIN" | "MEMBER",
-                    })
-                  }
-                  className="mt-1 block w-full rounded-lg border border-gray-300 px-3 py-2 shadow-sm focus:border-blue-500 focus:outline-none focus:ring-1 focus:ring-blue-500"
-                >
-                  <option value="MEMBER">Member</option>
-                  <option value="ADMIN">Admin</option>
-                </select>
-              </div>
-
-              {createUserError && (
-                <div className="rounded-md border border-red-200 bg-red-50 p-3">
-                  <p className="text-sm text-red-700">{createUserError}</p>
-                </div>
-              )}
-
-              <div className="rounded-md bg-yellow-50 p-3">
-                <p className="text-sm text-yellow-700">
-                  <strong>Default Password:</strong> Welcome123!
-                </p>
-                <p className="mt-1 text-sm text-yellow-600">
-                  Please share this password with the user manually.
-                </p>
-              </div>
-
-              <div className="flex space-x-3">
-                <button
-                  type="submit"
-                  disabled={isCreatingUser}
-                  className="flex-1 rounded-lg bg-blue-600 px-4 py-2 text-sm font-medium text-white shadow-sm transition-colors hover:bg-blue-700 disabled:opacity-50"
-                >
-                  {isCreatingUser ? "Creating..." : "Create User"}
-                </button>
-                <button
-                  type="button"
-                  onClick={handleCloseCreateUserModal}
-                  className="flex-1 rounded-lg border border-gray-300 bg-white px-4 py-2 text-sm font-medium text-gray-700 shadow-sm transition-colors hover:bg-gray-50"
-                >
-                  Cancel
-                </button>
-              </div>
-            </form>
-          </div>
-        </div>
-      )}
-
-      {/* Created User Success Modal */}
-      {createdUser && (
-        <div className="fixed inset-0 z-50 flex items-center justify-center bg-black bg-opacity-50">
-          <div className="mx-4 max-w-md rounded-lg bg-white p-6 shadow-xl">
-            <div className="mb-4 flex items-center">
-              <div className="mr-3 rounded-full bg-green-100 p-2">
-                <UserIcon className="h-5 w-5 text-green-600" />
-              </div>
-              <h3 className="text-lg font-medium text-gray-900">
-                User Created Successfully
-              </h3>
-            </div>
-
-            <div className="mb-6 space-y-3">
-              <div className="rounded-md bg-green-50 p-3">
-                <p className="text-sm text-green-700">
-                  <strong>User:</strong> {createdUser.user.name}
-                </p>
-                <p className="text-sm text-green-700">
-                  <strong>Email:</strong> {createdUser.user.email}
-                </p>
-                <p className="text-sm text-green-700">
-                  <strong>Role:</strong> {createdUser.user.role}
-                </p>
-              </div>
-
-              <div className="rounded-md bg-yellow-50 p-3">
-                <p className="text-sm font-medium text-yellow-800">
-                  Default Password
-                </p>
-                <p className="mt-1 text-sm text-yellow-700">
-                  <strong>{createdUser.defaultPassword}</strong>
-                </p>
-                <p className="mt-1 text-sm text-yellow-600">
-                  Please share this password with the user manually.
-                </p>
-              </div>
-            </div>
-
-            <button
-              onClick={handleCloseCreateUserModal}
-              className="w-full rounded-lg bg-blue-600 px-4 py-2 text-sm font-medium text-white shadow-sm transition-colors hover:bg-blue-700"
-            >
-              Close
-            </button>
-          </div>
-        </div>
+        <CreateUserModal
+          isOpen={showCreateUserModal}
+          onClose={() => setShowCreateUserModal(false)}
+        />
       )}
 
       {/* Delete Account Confirmation Modal */}
